+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## Unreleased
 
+- Enhance authentication documentation and API endpoint. (`5824077`)
 - Enhance Docker setup and documentation. (`90e2a39`)